feat: 카카오지도 페이지 hook,index,type 분리
