+++ src/components/kakaoMap/type.ts
@@ -1,0 +1,4 @@
+export type ChatUserDataType = {
+  writeUserId: number;
+  requestUserId: number;
+};
